correcciones viernes 17 mar

+++ src/controllers/pedido.js
@@ -5,30 +5,16 @@ const {
   requerimiento_producto,
   producto,
   area,
+  unidad,
+  categoria,
 } = require("../../config/db");
 const {
   generarPdfRequerimiento,
 } = require("../utills/PDF/requerimientoBienes");
 const _ = require("lodash");
 
-
-const getPedidoId = async(req, res, next ) => {
-
+const getPedidoId = async (req, res, next) => {
   try {
-    const get = await pedido.findAll()
-    res.status(200).json({ data: get });
-    next();
-    
-  } catch (error) {
-    console.log(error);
-    res.status(500).json();
-  }
-
-}
-
-const getPedido = async (req, res, next) => {
-  try {
-    const getArea = await area.findAll()
     const get = await pedido.findAll({
       include: [
         {
@@ -39,7 +25,65 @@ const getPedido = async (req, res, next) => {
               include: [
                 {
                   model: requerimiento_producto,
-                  include: [{ model: producto , attributes: {exclude:["categoria_id"]} }],
+                  include: [
+                    {
+                      model: producto,
+                      attributes: { exclude: ["categoria_id"] },
+                      include: [{ model: unidad }, { model: categoria }],
+                    },
+                  ],
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    });
+    const formatData = get
+      .map((item) => {
+        return {
+          id: item.id,
+          estado_pedido: item.estado,
+
+          requerimiento_pedidos: item.requerimiento_pedidos.map((data) => {
+            return {
+              requerimiento_id: data.requerimiento_id,
+              estado: data.estado,
+              almacen_id: data.requerimiento.almacen_id,
+            };
+          }),
+        };
+      })
+      .sort((a, b) => {
+        return b.id - a.id;
+      });
+    res.status(200).json({ data: formatData });
+    next();
+  } catch (error) {
+    console.log(error);
+    res.status(500).json();
+  }
+};
+
+const getPedido = async (req, res, next) => {
+  try {
+    const get = await pedido.findAll({
+      include: [
+        {
+          model: requerimiento_pedido,
+          include: [
+            {
+              model: requerimiento,
+              include: [
+                {
+                  model: requerimiento_producto,
+                  include: [
+                    {
+                      model: producto,
+                      attributes: { exclude: ["categoria_id"] },
+                      include: [{ model: unidad }, { model: categoria }],
+                    },
+                  ],
                 },
               ],
             },
@@ -48,36 +92,54 @@ const getPedido = async (req, res, next) => {
       ],
     });
 
-    const formatData = get.map((item) => {
-      return {
-        id: item.id,
-        fecha: item.fecha,
-        estado: item.estado,
-        area: (getArea.filter(dat => dat.id == item.area).map(item => item.nombre)).toString(),
-        celular: item.celular,
-        proyecto: item.proyecto,
-        solicitante: item.solicitante,
-        requerimiento_pedidos: item.requerimiento_pedidos
-        // producto: _.groupBy([
-        //   ...new Set(
-        //     item.requerimiento_pedidos
-        //       .map((data) =>
-        //         data.requerimiento.requerimiento_productos.map((dat) => {
-        //           return {
-        //             // nombre: dat.producto.nombre,
-        //             categoria: dat.producto.categoria,
-        //             cantidad: item.cantidad
-        //           };
-        //         })
-        //       )
-        //       .flat()
-        //   ),
-        // ], "categoria"),
-      };
-    });
+    const formatData = get
+      .map((item) => {
+        return {
+          id: item.id,
+          fecha: item.fecha,
+          estado: item.estado,
+          area: item.area,
+          celular: item.celular,
+          proyecto: item.proyecto,
+          solicitante: item.solicitante,
+          requerimiento_pedidos: item.requerimiento_pedidos.map((data) => {
+            return {
+              id: data.id,
+              requerimiento_id: data.requerimiento_id,
+              pedido_id: data.pedido_id,
+              estado: data.estado,
+              requerimiento: data.requerimiento,
+              productos: data.requerimiento.requerimiento_productos,
+            };
+          }),
+          productos: item.requerimiento_pedidos
+            .map((item) => item.requerimiento)
+            .map((dat) => dat.requerimiento_productos)
+            .flat()
+            .reduce((acc, value) => {
+              const product = acc.find(
+                (ele) => ele.producto_id === value.producto_id
+              );
+              if (product) {
+                product.cantidad =
+                  parseInt(product.cantidad) + parseInt(value.cantidad);
+              } else {
+                acc.push({
+                  producto_id: value.producto_id,
+                  cantidad: parseInt(value.cantidad),
+                  producto: value.producto,
+                });
+              }
+              return acc;
+            }, []),
+        };
+      })
+      .sort((a, b) => {
+        return b.id - a.id;
+      });
 
     //corregir los productos por categoria
-    res.status(200).json({ data: get });
+    res.status(200).json({ data: formatData });
     next();
   } catch (error) {
     console.log(error);
@@ -85,28 +147,13 @@ const getPedido = async (req, res, next) => {
   }
 };
 
-// const getPedidoById = async (req, res, next) => {
-//   let id = req.params.id;
-
-//   try {
-//     const getById = await pedido.findAll({
-//       where: { id: id },
-//     });
-//     res.status(200).json({ data: getById });
-//     next();
-//   } catch (error) {
-//     console.log(error);
-//     res.status(500).json({ error: error });
-//   }
-// };
-
 const postPedido = async (req, res, next) => {
   let info = {
     fecha: req.body.fecha,
     estado: req.body.estado,
     area: req.body.area,
     celular: req.body.celular,
-    proyecto: req.body.fecha,
+    proyecto: req.body.proyecto,
     solicitante: req.body.solicitante,
   };
   console.log(req.body.req_id);
@@ -125,14 +172,14 @@ const postPedido = async (req, res, next) => {
     );
 
     const updateRequerimiento = await requerimiento.update(
-      { completado: "Pedido" },
+      { estado: "Pedido", completado: "Pedido" },
       {
         where: {
           id: req.body.req_id,
         },
       }
     );
-    res.status(200).json({ msg: "Pedido generado con éxito!", status: 200 });
+    res.status(200).json({ msg: "Pedido realizado con éxito!", status: 200 });
 
     next();
   } catch (error) {
@@ -154,11 +201,44 @@ const updatePedido = async (req, res, next) => {
 
 const deletePedido = async (req, res, next) => {
   let id = req.params.id;
+  let data = req.query.ids;
+  const arr = data?.split(",");
+
+  console.log(arr);
   try {
-    let camp = await pedido.destroy({ where: { id: id } });
-    res.status(200).json({ msg: "Pedido eliminado con éxito!", status: 200 });
+    const getReqPedido = await requerimiento_pedido.findAll({
+      where: { pedido_id: id },
+    });
+
+    console.log(getReqPedido);
+    const filter = getReqPedido.filter(
+      (item) => item.estado === "Entregado" || item.estado === "En almacén"
+    );
+
+    if (filter.length > 0) {
+      return res.status(500).json({
+        msg: "No se pudo eliminar el pedido, debido a que algunos productos ya estan en almacén.",
+        status: 500,
+      });
+    } else {
+      let delReqPedido = await requerimiento_pedido.destroy({
+        where: { pedido_id: id },
+      });
+      let camp = await pedido.destroy({ where: { id: id } });
+
+      let requerimientoEstado = await requerimiento.update(
+        { estado: "Aprobado" },
+        { where: { id: arr } }
+      );
+
+      return res
+        .status(200)
+        .json({ msg: "Pedido eliminado con éxito!", status: 200 });
+    }
+
     next();
   } catch (error) {
+    console.log(error);
     res.status(500).json({ msg: "No se pudo eliminar.", status: 500 });
   }
 };
@@ -177,7 +257,13 @@ const getPedidoProducto = async (req, res, next) => {
               include: [
                 {
                   model: requerimiento_producto,
-                  include: [{ model: producto, attributes: { exclude: ["categoria_id"] }  }],
+                  include: [
+                    {
+                      model: producto,
+                      attributes: { exclude: ["categoria_id"] },
+                      include: [{ model: unidad }],
+                    },
+                  ],
                 },
               ],
             },
@@ -236,5 +322,5 @@ module.exports = {
   deletePedido,
   getPedidoProducto,
   descargarPedido,
-  getPedidoId
+  getPedidoId,
 };
