+++ src/controllers/requerimiento.js
@@ -1,50 +1,57 @@
+const dayjs = require("dayjs");
+const { where } = require("sequelize");
 const {
   requerimiento,
   requerimiento_producto,
   producto,
   area,
   almacen,
+  unidad,
 } = require("../../config/db");
 
 const getRequerimiento = async (req, res, next) => {
   try {
-
-    const getArea = await area.findAll()
-    const getAlmacen = await almacen.findAll()
     const get = await requerimiento.findAll({
       include: [
+        { model: almacen },
         {
           model: requerimiento_producto,
           include: [
-            { model: producto, attributes: { exclude: ["categoria_id"] } },
+            {
+              model: producto,
+              attributes: { exclude: ["categoria_id"] },
+              include: [{ model: unidad }],
+            },
           ],
         },
       ],
     });
 
-    const formatData = get.map(item => {
-
-      return{
-
-        id: item.id,
-        fecha_pedido: item.fecha_pedido,
-        fecha_entrega: item.fecha_entrega,
-        solicitante: item.solicitante,
-        area: (getArea.filter(dat => dat.id == item.area).map(item => item.nombre)).toString(),
-        celular: item.celular,
-        proyecto: item.proyecto,
-        codigo_requerimiento: item.codigo_requerimiento,
-        almacen_id: item.almacen_id,
-        almacen: (getAlmacen.filter(dat => dat.id == item.almacen_id).map(item => item.nombre)).toString(),
-        estado: item.estado,
-        aprobacion_jefe: item.aprobacion_jefe,
-        aprobacion_gerente: item.aprobacion_gerente,
-        aprobacion_superintendente: item.aprobacion_superintendente,
-        completado: item.completado,
-        requerimiento_productos: item.requerimiento_productos
-
-      }
-    })
+    const formatData = get
+      .map((item) => {
+        return {
+          id: item?.id,
+          fecha_pedido: dayjs(item?.fecha_pedido).format("YYYY-MM-DD"),
+          fecha_entrega: dayjs(item?.fecha_entrega).format("YYYY-MM-DD"),
+          solicitante: item?.solicitante,
+          area: item?.area,
+          celular: item?.celular,
+          proyecto: item?.proyecto,
+          codigo_requerimiento: item?.codigo_requerimiento,
+          almacen_id: item?.almacen_id,
+          almacen: item?.almacen?.nombre,
+          estado: item?.estado,
+          aprobacion_jefe: item?.aprobacion_jefe,
+          aprobacion_gerente: item?.aprobacion_gerente,
+          aprobacion_superintendente: item?.aprobacion_superintendente,
+          completado: item?.completado,
+          requerimiento_productos: item?.requerimiento_productos,
+          dni: item?.dni,
+        };
+      })
+      .sort((a, b) => {
+        return b.id - a.id;
+      });
     res.status(200).json({ data: formatData });
     next();
   } catch (error) {
@@ -71,6 +78,7 @@ const getRequerimientoById = async (req, res, next) => {
 const postARequerimiento = async (req, res, next) => {
   let data = req?.body?.map((item) => {
     return {
+      dni: item.dni,
       codigo: item.codigo,
       fecha_pedido: item.fecha_pedido,
       fecha_entrega: item.fecha_entrega,
@@ -79,12 +87,13 @@ const postARequerimiento = async (req, res, next) => {
       celular: item.celular,
       proyecto: item.proyecto,
       almacen_id: item.almacen_id,
-      completado: "Pendiente"
+      completado: "Pendiente",
+      estado: "Pendiente",
     };
   });
 
   try {
-    const post = await requerimiento.create(data[data.length - 1]);
+    const post = await requerimiento.create(data.at(-1));
 
     const productoRequerimiento = req.body.map((item) => {
       return {
@@ -108,7 +117,6 @@ const postARequerimiento = async (req, res, next) => {
   }
 };
 
-
 const updateRequerimiento = async (req, res, next) => {
   let id = req.params.id;
 
@@ -118,24 +126,32 @@ const updateRequerimiento = async (req, res, next) => {
       where: { id: id },
     });
 
-    if (
-      updateEstado.aprobacion_jefe &&
-      updateEstado.aprobacion_gerente &&
-      updateEstado.aprobacion_superintendente
-    ) {
-      let update = await requerimiento.update(
-        { estado: 1, completado: "Aprobado" },
-        { where: { id: id } }
-      );
+    if (updateEstado.estado === "Pendiente") {
+      if (
+        updateEstado.aprobacion_jefe &&
+        updateEstado.aprobacion_gerente &&
+        updateEstado.aprobacion_superintendente
+      ) {
+        let update = await requerimiento.update(
+          { estado: "Aprobado", completado: "Aprobado" },
+          { where: { id: id } }
+        );
+      } else {
+        let update = await requerimiento.update(
+          { estado: "Pendiente", completado: "Aprobado" },
+          { where: { id: id } }
+        );
+      }
+      return res
+        .status(200)
+        .json({ msg: "Requerimiento actualizado con éxito!", status: 200 });
     } else {
-      let update = await requerimiento.update(
-        { estado: 0, completado: "Aprobado" },
-        { where: { id: id } }
-      );
+      return res.status(200).json({
+        msg: "El requerimiento ya fue aprobado, no se puede actualizar.",
+        status: 200,
+      });
     }
-    res
-      .status(200)
-      .json({ msg: "Requerimiento actualizado con éxito!", status: 200 });
+
     next();
   } catch (error) {
     res.status(500).json({ msg: "No se pudo actualizar.", status: 500 });
@@ -144,40 +160,53 @@ const updateRequerimiento = async (req, res, next) => {
 
 const updateRequerimientoProducto = async (req, res, next) => {
   let id = req.params.id;
-
-  let updateCantidad = req.body.map((item) => {
-    return {
-      id: item.codigo_producto,
-      cantidad: item.cantidad,
-    };
-  });
-
-
+  const format = req.body
+    .map((item) => {
+      return {
+        area: item.area,
+        celular: item.celular,
+        dni: item.dni,
+        fecha_entrega: item.fecha_entrega,
+        fecha_pedido: item.fecha_pedido,
+        proyecto: item.proyecto,
+        solicitante: item.solicitante,
+        dni: item.dni,
+      };
+    })
+    .at(-1);
+    console.log(req.body);
   try {
-    // let update = await requerimiento.update(req.body, { where: { id: id } });
+    let update = await requerimiento.update(format, {
+      where: { id: id },
+    });
 
-    // let updateProducto = await requerimiento_producto.update(updateCantidad, {
-    //   where: { producto_id: updateIds },
-    // });
+    const updateRequerimientoProducto = req.body.map((item) => {
+      return {
+        requerimiento_id: id,
+        producto_id: item.codigo_producto,
+        cantidad: item.cantidad,
+      };
+    });
 
-    const updateMultiple = await Promise.all(
-      updateCantidad.map(
-        async (item) =>
-          await requerimiento_producto.update(
-            { cantidad: item.cantidad },
-            {
-              where: { producto_id: item.id, requerimiento_id: id },
-            }
-          )
-      )
+    console.log(updateRequerimientoProducto);
+
+    const delReqProducto = await requerimiento_producto.destroy({
+      where: { requerimiento_id: id },
+    });
+
+    let updateProducto = await requerimiento_producto.bulkCreate(
+      updateRequerimientoProducto,
+      {
+        ignoreDuplicates: false,
+      }
     );
-
 
     res
       .status(200)
       .json({ msg: "Requerimiento actualizado con éxito!", status: 200 });
     next();
   } catch (error) {
+    console.log(error);
     res.status(500).json({ msg: "No se pudo actualizar.", status: 500 });
   }
 };
@@ -185,13 +214,40 @@ const updateRequerimientoProducto = async (req, res, next) => {
 const deleteRequerimiento = async (req, res, next) => {
   let id = req.params.id;
   try {
-    let camp = await requerimiento.destroy({ where: { id: id } });
-    res
-      .status(200)
-      .json({ msg: "Requerimiento eliminado con éxito!", status: 200 });
+    let getReq = await requerimiento.findOne({ where: { id: id } });
+
+    console.log(getReq);
+    if (getReq.estado === "Pendiente" ) {
+      let delReqProducto = await requerimiento_producto.destroy({
+        where: { requerimiento_id: id },
+      });
+      let camp = await requerimiento.destroy({ where: { id: id } });
+      return res
+        .status(200)
+        .json({ msg: "Requerimiento eliminado con éxito!", status: 200 });
+    } else {
+      return res
+        .status(500)
+        .json({ msg: "No se puede eliminar el requerimiento.", status: 500 });
+    }
     next();
   } catch (error) {
     res.status(500).json({ msg: "No se pudo eliminar.", status: 500 });
+  }
+};
+
+const getLastId = async (req, res, next) => {
+  try {
+    const get = await requerimiento.findOne({
+      order: [["id", "DESC"]],
+    });
+
+    const getId = get ? get?.id + 1 : 1;
+    res.status(200).json({ data: getId });
+    next();
+  } catch (error) {
+    console.log(error);
+    res.status(500).json({ msg: "No se pudo obtener", status: 500 });
   }
 };
 
@@ -202,4 +258,5 @@ module.exports = {
   updateRequerimiento,
   deleteRequerimiento,
   updateRequerimientoProducto,
+  getLastId,
 };
